sidecar/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
numpy==1.25.2
llama-cpp-python==0.2.20